sched: drop group_capacity to 1 only if local group has extra capacity
Resending this patch since the original patch was munged. Thanks to Mike
Galbraith for pointing this out.

When SD_PREFER_SIBLING is set on a sched domain, drop group_capacity to 1
only if the local group has extra capacity. For niced task balancing, we pull
low weight tasks away from a sched group as long as there is capacity in other
groups. When all other groups are saturated, we do not drop capacity of the
niced group down to 1. This prevents active balance from kicking out the low
weight threads and which hurts system utilization.

Signed-off-by: Nikhil Rao <ncrao@google.com>